Completing project

+++ dev/hdr/trunc.h
@@ -6,19 +6,18 @@ class Trunc
     private:
         unsigned int N;
         double Q;
-        double m_max;
-
+        int m_max;
 
     public:
         Trunc(unsigned int N, double Q);
 
         double mu(double i);
 
-        double calc_mMax();
+        int calc_mMax();
 
-        double calc_nMax(int m);
+        int calc_nMax(int m);
 
-        double calc_nzMax(int m, int n);
+        int calc_nzMax(int m, int n);
 };
 
 #endif // TRUNC_H
